refactor: extract dashboard photo timeline hud
