+++ apps/ios/LogYourBody/Views/DashboardViewLiquid.swift
@@ -69,27 +69,27 @@ struct DashboardViewLiquid: View {
 
     // Timeline mode
     @State private var timelineMode: TimelineMode = .photo
-    @State private var photoDisplayMode: DashboardDisplayMode = .photo
+    @State var photoDisplayMode: DashboardDisplayMode = .photo
     @AppStorage(Constants.defaultHomeModeKey)
-    private var defaultHomeModeRawValue = DefaultHomeMode.default.rawValue
+    var defaultHomeModeRawValue = DefaultHomeMode.default.rawValue
 
     // Navigation state
     @State private var selectedTab: DashboardTab = .home
     @State private var isPhotosTabEnabled = true
-    @State private var selectedPhotoTimelineRootPage: PhotoTimelineRootPage = .timeline
+    @State var selectedPhotoTimelineRootPage: PhotoTimelineRootPage = .timeline
     @State var isMetricDetailActive = false
     @State var selectedMetricType: MetricType = .weight
     @State private var isStatsDestinationActive = false
-    @State private var isProgressPhotoAttachPresented = false
-    @State private var progressPhotoAttachTarget: BodyMetrics?
+    @State var isProgressPhotoAttachPresented = false
+    @State var progressPhotoAttachTarget: BodyMetrics?
     @State private var chartMode: ChartMode = .trend
     @State var bodyScoreRefreshToken = UUID()
     @State var isBodyScoreSharePresented = false
     @State var bodyScoreSharePayload: BodyScoreSharePayload?
     @State private var hasPerformedInitialRefresh = false
-    @State private var featureGateRefreshToken = UUID()
-    @State private var addEntryInitialTab = 0
-    @State private var addEntryIncludesGlp1Entry = false
+    @State var featureGateRefreshToken = UUID()
+    @State var addEntryInitialTab = 0
+    @State var addEntryIncludesGlp1Entry = false
 
     init(layoutMode: LayoutMode = .legacyTabbed) {
         self.layoutMode = layoutMode
@@ -371,7 +371,7 @@ struct DashboardViewLiquid: View {
         }
     }
 
-    private func refreshGlobalTimelineStore() {
+    func refreshGlobalTimelineStore() {
         let heightInches = convertHeightToInches(
             height: authManager.currentUser?.profile?.height,
             heightUnit: authManager.currentUser?.profile?.heightUnit
@@ -397,15 +397,15 @@ struct DashboardViewLiquid: View {
         }
     }
 
-    private var selectedDefaultHomeMode: DefaultHomeMode {
+    var selectedDefaultHomeMode: DefaultHomeMode {
         DefaultHomeMode(storedValue: defaultHomeModeRawValue)
     }
 
-    private var selectedHomeTimelineMode: TimelineMode {
+    var selectedHomeTimelineMode: TimelineMode {
         selectedDefaultHomeMode.timelineMode
     }
 
-    private var homeTimelineModeBinding: Binding<TimelineMode> {
+    var homeTimelineModeBinding: Binding<TimelineMode> {
         Binding(
             get: { selectedHomeTimelineMode },
             set: { newValue in
@@ -414,7 +414,7 @@ struct DashboardViewLiquid: View {
         )
     }
 
-    private var homeModeSwitch: some View {
+    var homeModeSwitch: some View {
         HStack(spacing: 4) {
             ForEach(DefaultHomeMode.allCases) { mode in
                 homeModeButton(mode)
@@ -433,7 +433,7 @@ struct DashboardViewLiquid: View {
         .accessibilityIdentifier("home_mode_switch")
     }
 
-    private func homeModeButton(_ mode: DefaultHomeMode) -> some View {
+    func homeModeButton(_ mode: DefaultHomeMode) -> some View {
         let isSelected = selectedDefaultHomeMode == mode
 
         return Button {
@@ -456,7 +456,7 @@ struct DashboardViewLiquid: View {
         .accessibilityIdentifier("home_mode_\(mode.rawValue)_button")
     }
 
-    private func homeTimelineHero(metric: BodyMetrics) -> some View {
+    func homeTimelineHero(metric: BodyMetrics) -> some View {
         let bodyScore = bodyScoreText()
 
         return DashboardHomeTimelineHero(
@@ -566,840 +566,6 @@ struct DashboardViewLiquid: View {
         )
     }
 
-    private var photoTimelineHUD: some View {
-        ScrollView(showsIndicators: false) {
-            VStack(spacing: 16) {
-                compactHeader
-                    .padding(.horizontal, 20)
-                    .padding(.top, 8)
-
-                syncStatusBanner
-                    .padding(.horizontal, 20)
-
-                homeModeSwitch
-                    .padding(.horizontal, 20)
-
-                if let metric = currentMetric {
-                    homeTimelineHero(metric: metric)
-                }
-
-                hudTimelineSection
-                    .padding(.horizontal, 20)
-
-                if isGlp1WeeklyCheckInEnabled {
-                    hudGlp1WeeklyCheckIn
-                        .padding(.horizontal, 20)
-                }
-
-                if isPhaseInsightEnabled {
-                    hudPhaseInsight
-                        .padding(.horizontal, 20)
-                }
-
-                hudStatsAction
-                    .padding(.horizontal, 20)
-
-                Spacer(minLength: 120)
-            }
-            .padding(.bottom, 24)
-        }
-        .refreshable {
-            await viewModel.refreshData(
-                authManager: authManager,
-                realtimeSyncManager: realtimeSyncManager
-            )
-            refreshGlobalTimelineStore()
-            if !viewModel.bodyMetrics.isEmpty {
-                updateAnimatedValues(for: selectedIndex)
-            }
-        }
-        .accessibilityIdentifier("photo_timeline_hud")
-    }
-
-    private var photoTimelineRoot: some View {
-        TabView(selection: $selectedPhotoTimelineRootPage) {
-            Group {
-                if bodyMetrics.isEmpty {
-                    photoTimelineHUDEmptyState
-                } else {
-                    photoTimelineHUD
-                }
-            }
-            .tag(PhotoTimelineRootPage.timeline)
-            .accessibilityIdentifier("photo_timeline_root_page_timeline")
-
-            photoTimelineAnalyticsPage
-                .tag(PhotoTimelineRootPage.analytics)
-                .accessibilityIdentifier("photo_timeline_root_page_analytics")
-        }
-        .tabViewStyle(.page(indexDisplayMode: .always))
-        .accessibilityIdentifier("photo_timeline_root_pager")
-    }
-
-    private var hudPhotoStage: some View {
-        ZStack(alignment: .bottomLeading) {
-            ProgressPhotoCarouselView(
-                currentMetric: currentMetric,
-                historicalMetrics: bodyMetrics,
-                selectedMetricsIndex: $selectedIndex,
-                displayMode: $photoDisplayMode
-            )
-            .frame(height: 330)
-            .clipShape(RoundedRectangle(cornerRadius: 28))
-            .overlay(
-                RoundedRectangle(cornerRadius: 28)
-                    .stroke(Color.white.opacity(0.12), lineWidth: 1)
-            )
-
-            if !PhotoTimelineHUDPolicy.hasUsablePhoto(currentMetric) {
-                hudMissingPhotoOverlay
-            }
-
-            hudPhotoDateChip
-                .padding(14)
-        }
-        .accessibilityIdentifier("photo_timeline_hud_photo_stage")
-    }
-
-    private var hudMissingPhotoOverlay: some View {
-        VStack(alignment: .leading, spacing: 10) {
-            Image(systemName: "photo.on.rectangle.angled")
-                .font(.system(size: 28, weight: .semibold))
-                .foregroundColor(Color.white.opacity(0.82))
-
-            VStack(alignment: .leading, spacing: 4) {
-                Text("No progress photo")
-                    .font(.system(size: 18, weight: .semibold))
-                    .foregroundColor(Color.white)
-
-                Text("This timeline point has metrics only.")
-                    .font(.system(size: 13, weight: .medium))
-                    .foregroundColor(Color.white.opacity(0.68))
-            }
-
-            Button {
-                presentProgressPhotoAttach(for: progressPhotoAttachMetric)
-            } label: {
-                Label("Add photo", systemImage: "camera.fill")
-                    .font(.system(size: 13, weight: .semibold))
-                    .padding(.horizontal, 12)
-                    .padding(.vertical, 8)
-                    .background(Capsule().fill(Color.white.opacity(0.14)))
-                    .foregroundColor(.white)
-            }
-            .buttonStyle(.plain)
-            .accessibilityIdentifier("photo_timeline_hud_add_photo_button")
-        }
-        .padding(18)
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .bottomLeading)
-        .background(
-            LinearGradient(
-                colors: [
-                    Color.black.opacity(0.05),
-                    Color.black.opacity(0.58)
-                ],
-                startPoint: .top,
-                endPoint: .bottom
-            )
-            .clipShape(RoundedRectangle(cornerRadius: 28))
-        )
-    }
-
-    private var hudPhotoDateChip: some View {
-        Text(formatHUDDate(currentMetric?.date ?? Date()))
-            .font(.system(size: 12, weight: .semibold))
-            .foregroundColor(.white)
-            .padding(.horizontal, 10)
-            .padding(.vertical, 6)
-            .background(Capsule().fill(Color.black.opacity(0.45)))
-            .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topTrailing)
-            .accessibilityIdentifier("photo_timeline_hud_date_chip")
-    }
-
-    private var hudTimelineSection: some View {
-        VStack(spacing: 12) {
-            if !globalTimelineStore.weeklyBuckets.isEmpty ||
-                !globalTimelineStore.monthlyBuckets.isEmpty ||
-                !globalTimelineStore.yearlyBuckets.isEmpty {
-                GlobalTimelineHeader(
-                    weeklyBuckets: globalTimelineStore.weeklyBuckets,
-                    monthlyBuckets: globalTimelineStore.monthlyBuckets,
-                    yearlyBuckets: globalTimelineStore.yearlyBuckets,
-                    cursor: globalTimelineStore.cursor,
-                    onCursorChange: { cursor in
-                        globalTimelineStore.updateCursor(cursor)
-                        selectClosestMetric(to: cursor.date)
-                    },
-                    onTodayTap: {
-                        globalTimelineStore.selectToday()
-                        if let cursor = globalTimelineStore.cursor {
-                            selectClosestMetric(to: cursor.date)
-                        }
-                    }
-                )
-            }
-
-            DashboardTimelineScrubber(
-                bodyMetrics: bodyMetrics,
-                selectedIndex: $selectedIndex,
-                timelineMode: homeTimelineModeBinding
-            )
-        }
-        .padding(.vertical, 10)
-        .background(
-            RoundedRectangle(cornerRadius: 20)
-                .fill(Color.white.opacity(0.06))
-        )
-        .accessibilityIdentifier("photo_timeline_hud_timeline")
-    }
-
-    private var hudMetricsStrip: some View {
-        let bucket = activeTimelineBucket
-        let weight = bucket?.metrics.weight ?? missingHUDMetric
-        let bodyFat = bucket?.metrics.bodyFat ?? missingHUDMetric
-        let ffmi = bucket?.metrics.ffmi ?? missingHUDMetric
-        let steps = bucket?.metrics.steps ?? missingHUDMetric
-
-        return LazyVGrid(
-            columns: [
-                GridItem(.flexible(), spacing: 10),
-                GridItem(.flexible(), spacing: 10)
-            ],
-            spacing: 10
-        ) {
-            hudMetricTile(
-                title: "Weight",
-                value: formattedHUDWeight(weight.value),
-                caption: PhotoTimelineHUDPolicy.stateText(
-                    presence: weight.presence,
-                    confidence: weight.confidence
-                ),
-                icon: "scalemass.fill",
-                color: Color.metricAccentWeight,
-                onTap: {
-                    selectedMetricType = .weight
-                    isMetricDetailActive = true
-                }
-            )
-
-            hudMetricTile(
-                title: "Body Fat",
-                value: formattedHUDBodyFat(bodyFat.value),
-                caption: PhotoTimelineHUDPolicy.stateText(
-                    presence: bodyFat.presence,
-                    confidence: bodyFat.confidence
-                ),
-                icon: "percent",
-                color: Color.metricAccentBodyFat,
-                onTap: {
-                    selectedMetricType = .bodyFat
-                    isMetricDetailActive = true
-                }
-            )
-
-            hudMetricTile(
-                title: "FFMI",
-                value: formattedHUDFFMI(ffmi.value),
-                caption: PhotoTimelineHUDPolicy.stateText(
-                    presence: ffmi.presence,
-                    confidence: ffmi.confidence
-                ),
-                icon: "figure.arms.open",
-                color: Color.metricAccentFFMI,
-                onTap: {
-                    selectedMetricType = .ffmi
-                    isMetricDetailActive = true
-                }
-            )
-
-            hudMetricTile(
-                title: "Steps",
-                value: formattedHUDSteps(steps.value),
-                caption: PhotoTimelineHUDPolicy.stateText(
-                    presence: steps.presence,
-                    confidence: steps.confidence
-                ),
-                icon: "flame.fill",
-                color: Color.metricAccentSteps,
-                onTap: {
-                    selectedMetricType = .steps
-                    isMetricDetailActive = true
-                }
-            )
-        }
-        .accessibilityIdentifier("photo_timeline_hud_metrics")
-    }
-
-    private var hudPhaseInsight: some View {
-        let insight = PhaseInsightPolicy.insight(for: bodyMetrics)
-
-        return HStack(alignment: .top, spacing: 12) {
-            Image(systemName: phaseInsightIcon(for: insight.kind))
-                .font(.system(size: 15, weight: .semibold))
-                .foregroundColor(phaseInsightColor(for: insight.kind))
-                .frame(width: 34, height: 34)
-                .background(
-                    Circle()
-                        .fill(phaseInsightColor(for: insight.kind).opacity(0.15))
-                )
-
-            VStack(alignment: .leading, spacing: 5) {
-                HStack(spacing: 8) {
-                    Text(insight.title)
-                        .font(.system(size: 15, weight: .semibold))
-                        .foregroundColor(Color.liquidTextPrimary)
-
-                    if let delta = insight.weightDeltaPercentPerWeek {
-                        Text(formatPhaseInsightDelta(delta))
-                            .font(.system(size: 11, weight: .bold))
-                            .monospacedDigit()
-                            .foregroundColor(phaseInsightColor(for: insight.kind))
-                            .padding(.horizontal, 7)
-                            .padding(.vertical, 4)
-                            .background(
-                                Capsule()
-                                    .fill(phaseInsightColor(for: insight.kind).opacity(0.14))
-                            )
-                    }
-                }
-
-                Text(insight.message)
-                    .font(.system(size: 12, weight: .medium))
-                    .foregroundColor(Color.liquidTextPrimary.opacity(0.68))
-                    .fixedSize(horizontal: false, vertical: true)
-
-                if let detail = insight.detail {
-                    Text(detail)
-                        .font(.system(size: 11, weight: .medium))
-                        .foregroundColor(Color.liquidTextPrimary.opacity(0.52))
-                        .fixedSize(horizontal: false, vertical: true)
-                }
-            }
-
-            Spacer(minLength: 0)
-        }
-        .padding(14)
-        .background(
-            RoundedRectangle(cornerRadius: 18)
-                .fill(Color.white.opacity(0.06))
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: 18)
-                .stroke(Color.white.opacity(0.08), lineWidth: 1)
-        )
-        .accessibilityElement(children: .combine)
-        .accessibilityIdentifier("photo_timeline_hud_phase_insight")
-        .accessibilityLabel(phaseInsightAccessibilityLabel(for: insight))
-    }
-
-    private var hudGlp1WeeklyCheckIn: some View {
-        let summary = Glp1WeeklyCheckInPolicy.summary(
-            medications: glp1Medications,
-            doseLogs: glp1DoseLogs
-        )
-
-        return Button {
-            presentAddEntrySheet(initialTab: 3, includesGlp1Entry: true)
-        } label: {
-            HStack(alignment: .center, spacing: 12) {
-                Image(systemName: "syringe")
-                    .font(.system(size: 15, weight: .semibold))
-                    .foregroundColor(glp1WeeklyCheckInColor(for: summary.status))
-                    .frame(width: 34, height: 34)
-                    .background(
-                        Circle()
-                            .fill(glp1WeeklyCheckInColor(for: summary.status).opacity(0.15))
-                    )
-
-                VStack(alignment: .leading, spacing: 5) {
-                    HStack(spacing: 8) {
-                        Text(summary.title)
-                            .font(.system(size: 15, weight: .semibold))
-                            .foregroundColor(Color.liquidTextPrimary)
-
-                        if let latestDoseText = summary.latestDoseText {
-                            Text(latestDoseText)
-                                .font(.system(size: 11, weight: .bold))
-                                .monospacedDigit()
-                                .foregroundColor(glp1WeeklyCheckInColor(for: summary.status))
-                                .padding(.horizontal, 7)
-                                .padding(.vertical, 4)
-                                .background(
-                                    Capsule()
-                                        .fill(glp1WeeklyCheckInColor(for: summary.status).opacity(0.14))
-                                )
-                        }
-                    }
-
-                    Text(summary.message)
-                        .font(.system(size: 12, weight: .medium))
-                        .foregroundColor(Color.liquidTextPrimary.opacity(0.68))
-                        .fixedSize(horizontal: false, vertical: true)
-                }
-
-                Spacer(minLength: 0)
-
-                Text(summary.actionTitle)
-                    .font(.system(size: 12, weight: .semibold))
-                    .foregroundColor(.black)
-                    .padding(.horizontal, 12)
-                    .padding(.vertical, 8)
-                    .background(Capsule().fill(Color.white))
-            }
-            .padding(14)
-            .background(
-                RoundedRectangle(cornerRadius: 18)
-                    .fill(Color.white.opacity(0.06))
-            )
-            .overlay(
-                RoundedRectangle(cornerRadius: 18)
-                    .stroke(Color.white.opacity(0.08), lineWidth: 1)
-            )
-        }
-        .buttonStyle(.plain)
-        .accessibilityIdentifier("photo_timeline_hud_glp1_weekly_checkin")
-        .accessibilityLabel("\(summary.title). \(summary.message). \(summary.actionTitle)")
-    }
-
-    private var hudStatsAction: some View {
-        Button {
-            HapticManager.shared.selection()
-            withAnimation(.easeInOut(duration: 0.25)) {
-                selectedPhotoTimelineRootPage = .analytics
-            }
-        } label: {
-            HStack(spacing: 12) {
-                Image(systemName: "chart.line.uptrend.xyaxis")
-                    .font(.system(size: 16, weight: .semibold))
-                    .foregroundColor(Color.metricAccent)
-                    .frame(width: 32, height: 32)
-                    .background(
-                        Circle()
-                            .fill(Color.white.opacity(0.08))
-                    )
-
-                VStack(alignment: .leading, spacing: 2) {
-                    Text("Stats")
-                        .font(.system(size: 15, weight: .semibold))
-                        .foregroundColor(Color.liquidTextPrimary)
-
-                    Text("Charts, sources, and history")
-                        .font(.system(size: 12, weight: .medium))
-                        .foregroundColor(Color.liquidTextPrimary.opacity(0.58))
-                }
-
-                Spacer()
-
-                Image(systemName: "chevron.right")
-                    .font(.system(size: 13, weight: .semibold))
-                    .foregroundColor(Color.liquidTextPrimary.opacity(0.42))
-            }
-            .padding(14)
-            .background(
-                RoundedRectangle(cornerRadius: 18)
-                    .fill(Color.white.opacity(0.06))
-            )
-            .overlay(
-                RoundedRectangle(cornerRadius: 18)
-                    .stroke(Color.white.opacity(0.08), lineWidth: 1)
-            )
-        }
-        .buttonStyle(.plain)
-        .accessibilityIdentifier("photo_timeline_hud_stats_button")
-        .accessibilityLabel("Open stats")
-    }
-
-    private var photoTimelineStatsDestination: some View {
-        photoTimelineAnalyticsPage
-            .navigationTitle("Stats")
-            .navigationBarTitleDisplayMode(.inline)
-            .accessibilityIdentifier("photo_timeline_stats_destination")
-    }
-
-    private var photoTimelineAnalyticsPage: some View {
-        ZStack {
-            Color.metricCanvas.ignoresSafeArea()
-
-            ScrollView(showsIndicators: false) {
-                VStack(alignment: .leading, spacing: 18) {
-                    photoTimelineStatsHeader
-                        .padding(.horizontal, 20)
-
-                    photoTimelinePresenceSummary
-                        .padding(.horizontal, 20)
-
-                    metricsView
-                }
-                .padding(.top, 14)
-                .padding(.bottom, 32)
-            }
-        }
-    }
-
-    private var photoTimelineStatsHeader: some View {
-        VStack(alignment: .leading, spacing: 6) {
-            Text("Body trends")
-                .font(.system(size: 28, weight: .bold))
-                .foregroundColor(.white)
-
-            Text("Open a metric for chart and history.")
-                .font(.system(size: 14, weight: .medium))
-                .foregroundColor(Color.metricTextSecondary)
-        }
-    }
-
-    private var photoTimelinePresenceSummary: some View {
-        VStack(alignment: .leading, spacing: 12) {
-            HStack {
-                Text("Timeline states")
-                    .font(.system(size: 15, weight: .semibold))
-                    .foregroundColor(.white)
-
-                Spacer()
-
-                Text("\(timelinePresenceValueCount) values")
-                    .font(.system(size: 12, weight: .semibold))
-                    .foregroundColor(Color.metricTextTertiary)
-            }
-
-            LazyVGrid(
-                columns: [
-                    GridItem(.flexible(), spacing: 8),
-                    GridItem(.flexible(), spacing: 8)
-                ],
-                spacing: 8
-            ) {
-                ForEach(MetricPresence.allCases, id: \.rawValue) { presence in
-                    photoTimelinePresenceChip(
-                        title: photoTimelinePresenceLabel(for: presence),
-                        count: timelinePresenceCounts[presence] ?? 0,
-                        color: photoTimelinePresenceColor(for: presence)
-                    )
-                }
-            }
-        }
-        .padding(14)
-        .background(
-            RoundedRectangle(cornerRadius: 18)
-                .fill(Color.white.opacity(0.06))
-        )
-        .overlay(
-            RoundedRectangle(cornerRadius: 18)
-                .stroke(Color.white.opacity(0.08), lineWidth: 1)
-        )
-        .accessibilityIdentifier("photo_timeline_stats_presence_summary")
-    }
-
-    private func photoTimelinePresenceChip(
-        title: String,
-        count: Int,
-        color: Color
-    ) -> some View {
-        HStack(spacing: 8) {
-            Circle()
-                .fill(color)
-                .frame(width: 8, height: 8)
-
-            Text(title)
-                .font(.system(size: 12, weight: .semibold))
-                .foregroundColor(Color.metricTextSecondary)
-
-            Spacer(minLength: 4)
-
-            Text("\(count)")
-                .font(.system(size: 12, weight: .bold))
-                .foregroundColor(.white)
-        }
-        .padding(.horizontal, 10)
-        .padding(.vertical, 9)
-        .background(
-            RoundedRectangle(cornerRadius: 12)
-                .fill(Color.white.opacity(0.055))
-        )
-        .accessibilityLabel("\(title), \(count) timeline values")
-    }
-
-    private func hudMetricTile(
-        title: String,
-        value: String,
-        caption: String,
-        icon: String,
-        color: Color,
-        onTap: @escaping () -> Void
-    ) -> some View {
-        Button(action: onTap) {
-            VStack(alignment: .leading, spacing: 10) {
-                HStack(spacing: 8) {
-                    Image(systemName: icon)
-                        .font(.system(size: 13, weight: .semibold))
-                        .foregroundColor(color)
-
-                    Text(title)
-                        .font(.system(size: 12, weight: .semibold))
-                        .foregroundColor(Color.liquidTextPrimary.opacity(0.68))
-
-                    Spacer(minLength: 0)
-                }
-
-                Text(value)
-                    .font(.system(size: 24, weight: .bold, design: .rounded))
-                    .foregroundColor(Color.liquidTextPrimary)
-                    .monospacedDigit()
-                    .minimumScaleFactor(0.72)
-                    .lineLimit(1)
-
-                Text(caption)
-                    .font(.system(size: 11, weight: .medium))
-                    .foregroundColor(Color.liquidTextPrimary.opacity(0.54))
-                    .lineLimit(1)
-                    .minimumScaleFactor(0.78)
-            }
-            .padding(14)
-            .frame(maxWidth: .infinity, minHeight: 116, alignment: .leading)
-            .background(
-                RoundedRectangle(cornerRadius: 18)
-                    .fill(Color.white.opacity(0.07))
-            )
-            .overlay(
-                RoundedRectangle(cornerRadius: 18)
-                    .stroke(Color.white.opacity(0.08), lineWidth: 1)
-            )
-        }
-        .buttonStyle(.plain)
-        .accessibilityLabel("\(title), \(value), \(caption)")
-    }
-
-    private var activeTimelineBucket: GlobalTimelineBucket? {
-        if let cursor = globalTimelineStore.cursor,
-           let bucket = globalTimelineStore.bucket(for: cursor) {
-            return bucket
-        }
-
-        return globalTimelineStore.weeklyBuckets.last
-    }
-
-    private var progressPhotoAttachFallbackDate: Date {
-        currentMetric?.date ?? globalTimelineStore.cursor?.date ?? activeTimelineBucket?.endDate ?? Date()
-    }
-
-    private var progressPhotoAttachMetric: BodyMetrics? {
-        if let currentMetric {
-            return currentMetric
-        }
-
-        let targetDate = progressPhotoAttachFallbackDate
-        return bodyMetrics.min { lhs, rhs in
-            abs(lhs.date.timeIntervalSince(targetDate)) < abs(rhs.date.timeIntervalSince(targetDate))
-        }
-    }
-
-    private func presentProgressPhotoAttach(for metric: BodyMetrics?) {
-        progressPhotoAttachTarget = metric
-        HapticManager.shared.selection()
-        isProgressPhotoAttachPresented = true
-    }
-
-    @MainActor
-    private func handleProgressPhotoAttachComplete() async {
-        await viewModel.refreshData(
-            authManager: authManager,
-            realtimeSyncManager: realtimeSyncManager
-        )
-        refreshGlobalTimelineStore()
-
-        if let targetId = progressPhotoAttachTarget?.id,
-           let refreshedIndex = bodyMetrics.firstIndex(where: { $0.id == targetId }) {
-            selectedIndex = refreshedIndex
-            updateAnimatedValues(for: refreshedIndex)
-        } else if !bodyMetrics.isEmpty {
-            selectedIndex = min(selectedIndex, bodyMetrics.count - 1)
-            updateAnimatedValues(for: selectedIndex)
-        }
-    }
-
-    private var timelinePresenceValues: [GlobalTimelineMetricValue] {
-        globalTimelineStore.weeklyBuckets.flatMap { bucket in
-            [
-                bucket.metrics.weight,
-                bucket.metrics.bodyFat,
-                bucket.metrics.ffmi,
-                bucket.metrics.steps
-            ]
-        }
-    }
-
-    private var timelinePresenceCounts: [MetricPresence: Int] {
-        timelinePresenceValues.reduce(into: [:]) { counts, value in
-            counts[value.presence, default: 0] += 1
-        }
-    }
-
-    private var timelinePresenceValueCount: Int {
-        timelinePresenceValues.count
-    }
-
-    private var missingHUDMetric: GlobalTimelineMetricValue {
-        GlobalTimelineMetricValue(value: nil, presence: .missing)
-    }
-
-    private var isPhaseInsightEnabled: Bool {
-        _ = featureGateRefreshToken
-
-        #if DEBUG
-        if ProcessInfo.processInfo.arguments.contains("-lybUITestPhaseInsightFixture") {
-            return true
-        }
-        #endif
-
-        return PhaseInsightPolicy.shouldShowPhaseInsight(
-            gateEnabled: AnalyticsService.shared.isFeatureEnabled(
-                flagKey: Constants.phaseInsightFlagKey
-            )
-        )
-    }
-
-    private var isGlp1WeeklyCheckInEnabled: Bool {
-        _ = featureGateRefreshToken
-
-        #if DEBUG
-        if ProcessInfo.processInfo.arguments.contains("-lybUITestGlp1WeeklyCheckInFixture") {
-            return true
-        }
-        #endif
-
-        return Glp1WeeklyCheckInPolicy.shouldShowWeeklyCheckIn(
-            gateEnabled: AnalyticsService.shared.isFeatureEnabled(
-                flagKey: Constants.glp1WeeklyCheckInFlagKey
-            )
-        )
-    }
-
-    private func loadGlp1WeeklyCheckInDataIfNeeded() {
-        guard isGlp1WeeklyCheckInEnabled else { return }
-
-        Task {
-            await loadGlp1WeeklyCheckInData()
-        }
-    }
-
-    private func presentAddEntrySheet(initialTab: Int = 0, includesGlp1Entry: Bool = false) {
-        addEntryInitialTab = initialTab
-        addEntryIncludesGlp1Entry = includesGlp1Entry
-        showAddEntrySheet = true
-    }
-
-    private func phaseInsightIcon(for kind: PhaseInsightKind) -> String {
-        switch kind {
-        case .cutting:
-            return "chart.line.downtrend.xyaxis"
-        case .maintaining:
-            return "equal.circle.fill"
-        case .gaining:
-            return "chart.line.uptrend.xyaxis"
-        case .insufficientData:
-            return "clock.badge.questionmark.fill"
-        }
-    }
-
-    private func phaseInsightColor(for kind: PhaseInsightKind) -> Color {
-        switch kind {
-        case .cutting:
-            return Color.metricAccentBodyFat
-        case .maintaining:
-            return Color.metricAccent
-        case .gaining:
-            return Color.metricAccentWeight
-        case .insufficientData:
-            return Color.metricTextTertiary
-        }
-    }
-
-    private func phaseInsightAccessibilityLabel(for insight: PhaseInsight) -> String {
-        [
-            insight.title,
-            insight.message,
-            insight.detail
-        ]
-        .compactMap { $0 }
-        .joined(separator: ". ")
-    }
-
-    private func glp1WeeklyCheckInColor(for status: Glp1WeeklyCheckInStatus) -> Color {
-        switch status {
-        case .setup:
-            return Color.metricAccent
-        case .due:
-            return Color.metricAccentWeight
-        case .logged:
-            return Color.metricAccentSteps
-        }
-    }
-
-    private func formatPhaseInsightDelta(_ value: Double) -> String {
-        String(format: "%+.1f%%/wk", value)
-    }
-
-    private func photoTimelinePresenceLabel(for presence: MetricPresence) -> String {
-        switch presence {
-        case .present:
-            return "Measured"
-        case .interpolated:
-            return "Interpolated"
-        case .lastKnown:
-            return "Last known"
-        case .missing:
-            return "Missing"
-        }
-    }
-
-    private func photoTimelinePresenceColor(for presence: MetricPresence) -> Color {
-        switch presence {
-        case .present:
-            return Color.metricChartLine
-        case .interpolated:
-            return Color.metricAccentBodyFat
-        case .lastKnown:
-            return Color.metricAccentFFMI
-        case .missing:
-            return Color.metricTextTertiary
-        }
-    }
-
-    private func selectClosestMetric(to date: Date) {
-        guard let match = bodyMetrics.enumerated().min(by: { lhs, rhs in
-            abs(lhs.element.date.timeIntervalSince(date)) < abs(rhs.element.date.timeIntervalSince(date))
-        }) else {
-            return
-        }
-
-        selectedIndex = match.offset
-        updateAnimatedValues(for: match.offset)
-    }
-
-    private func formatHUDDate(_ date: Date) -> String {
-        FormatterCache.mediumDateFormatter.string(from: date)
-    }
-
-    private func formattedHUDWeight(_ value: Double?) -> String {
-        guard let value else { return "–" }
-        return "\(formatWeightValue(value)) \(weightUnit)"
-    }
-
-    private func formattedHUDBodyFat(_ value: Double?) -> String {
-        guard let value else { return "–" }
-        return "\(formatBodyFatValue(value))%"
-    }
-
-    private func formattedHUDFFMI(_ value: Double?) -> String {
-        guard let value else { return "–" }
-        return String(format: "%.1f", value)
-    }
-
-    private func formattedHUDSteps(_ value: Double?) -> String {
-        guard let value else { return "–" }
-        return FormatterCache.stepsFormatter.string(from: NSNumber(value: value)) ?? "\(Int(value))"
-    }
-
     private func loadMetricsOrder() {
         guard !metricsOrderData.isEmpty else { return }
 
@@ -1417,7 +583,7 @@ struct DashboardViewLiquid: View {
 
     // MARK: - Metrics View
 
-    private var metricsView: some View {
+    var metricsView: some View {
         let stepsGoalText: String? = {
             let formatted = FormatterCache.stepsFormatter.string(from: NSNumber(value: stepGoal)) ?? "\(stepGoal)"
             return "Goal \(formatted) steps"
@@ -1656,61 +822,6 @@ struct DashboardViewLiquid: View {
         DashboardEmptyStateLiquid {
             showAddEntrySheet = true
         }
-    }
-
-    private var photoTimelineHUDEmptyState: some View {
-        VStack(spacing: 16) {
-            compactHeader
-                .padding(.horizontal, 20)
-                .padding(.top, 8)
-
-            syncStatusBanner
-                .padding(.horizontal, 20)
-
-            VStack(alignment: .leading, spacing: 16) {
-                Image(systemName: "camera.metering.center.weighted")
-                    .font(.system(size: 34, weight: .semibold))
-                    .foregroundColor(Color.white.opacity(0.78))
-
-                VStack(alignment: .leading, spacing: 6) {
-                    Text("Start with a photo")
-                        .font(.system(size: 28, weight: .bold))
-                        .foregroundColor(.white)
-
-                    Text("Add a progress photo or weight entry to build your body-composition timeline.")
-                        .font(.system(size: 15, weight: .medium))
-                        .foregroundColor(Color.white.opacity(0.68))
-                        .fixedSize(horizontal: false, vertical: true)
-                }
-
-                Button {
-                    presentProgressPhotoAttach(for: nil)
-                } label: {
-                    Label("Add photo", systemImage: "camera.fill")
-                        .font(.system(size: 15, weight: .semibold))
-                        .frame(maxWidth: .infinity)
-                        .padding(.vertical, 14)
-                        .background(Capsule().fill(Color.white))
-                        .foregroundColor(.black)
-                }
-                .buttonStyle(.plain)
-                .accessibilityIdentifier("photo_timeline_hud_empty_add_photo_button")
-            }
-            .padding(24)
-            .frame(maxWidth: .infinity, minHeight: 420, alignment: .bottomLeading)
-            .background(
-                RoundedRectangle(cornerRadius: 28)
-                    .fill(Color.white.opacity(0.07))
-            )
-            .overlay(
-                RoundedRectangle(cornerRadius: 28)
-                    .stroke(Color.white.opacity(0.10), lineWidth: 1)
-            )
-            .padding(.horizontal, 20)
-
-            Spacer(minLength: 80)
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .top)
     }
 }
